requirements.txt
streamlit>=1.31.0
requests>=2.27.0
google-generativeai==0.3.1
ipython==8.18.1
pyautogui==0.9.53